pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "transcriber-shell"
version = "0.1.0"
description = "Lineation (mask / Kraken / Glyph Machina) + PageXML checks + LLM transcription against the Academic Transcription Protocol"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "transcriber-shell contributors" }]
dependencies = [
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "pyyaml>=6.0",
  "anthropic>=0.40.0",
  "httpx>=0.27.0",
  "openai>=1.50.0",
  "playwright>=1.49.0",
  "numpy>=1.26.0",
  "pillow>=10.0.0",
  "tkinterdnd2>=0.3.0",
]

[project.optional-dependencies]
xml-xsd = ["lxml>=5.0"]
gemini = ["google-generativeai>=0.8.0"]
api = ["fastapi>=0.115.0", "uvicorn[standard]>=0.32.0", "python-multipart>=0.0.9"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24.0", "httpx>=0.27.0"]
mask = ["opencv-python-headless>=4.9.0", "scipy>=1.11.0", "torch>=2.0.0"]
kraken = ["kraken>=5.0.0", "torch>=2.0.0"]
viz = ["matplotlib>=3.8.0"]

[project.scripts]
transcriber-shell = "transcriber_shell.cli:main"
transcriber-shell-gui = "transcriber_shell.gui:main"

[tool.hatch.build.targets.sdist]
include = [
  "src/transcriber_shell",
  "README.md",
  "LICENSE",
  "docs",
  "ground_truth",
  "fixtures",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
